requirements.txt
streamlit
mistralai
python-dotenv
PyPDF2
